# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Portfolio manager API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/kaseat/pManager/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "host": "totallink.ru",
    "basePath": "/api",
    "paths": {
...
        "/securities": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Gets securities avaliable",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "securities"
                ],
                "summary": "Get securities",
                "operationId": "get-securities",
                "parameters": [
                    {
                        "enum": [
                            "none",
                            "ticker",
                            "isin",
                            "figi"
                        ],
                        "type": "string",
                        "default": "none",
                        "description": "Filter by",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter value",
                        "name": "by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns success status",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Instrument"
                            }
                        }
                    },
                    "400": {
                        "description": "Returns when any processing error occurs",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Returns when authentication error occurs",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Adds securities",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "securities"
                ],
                "summary": "Add securities",
                "operationId": "add-securities",
                "parameters": [
                    {
                        "description": "Instrument info",
                        "name": "instrument",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Instrument"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns success status",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Instrument"
                            }
                        }
                    },
                    "400": {
                        "description": "Returns when any processing error occurs",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Returns when authentication error occurs",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            }
        },
        "/securities/sync": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Sync intruments dimension",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "securities"
                ],
                "summary": "Sync securities",
                "operationId": "sync-securities",
                "responses": {
                    "200": {
                        "description": "Returns success status",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.commonResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Returns when any processing error occurs",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Returns when authentication error occurs",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.commonResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "api.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Something went wrong"
                }
            }
        },
...
        "models.Instrument": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string",
                    "example": "USD"
                },
                "figi": {
                    "type": "string",
                    "example": "BBG000HLJ7M4"
                },
                "isin": {
                    "type": "string",
                    "example": "US45867G1013"
                },
                "name": {
                    "type": "string",
                    "example": "InterDigItal Inc"
                },
                "priceUptdTime": {
                    "type": "string",
                    "example": "2020-06-06T15:54:05Z"
                },
                "ticker": {
                    "type": "string",
                    "example": "IDCC"
                },
                "type": {
                    "type": "string",
                    "example": "Stock"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}